try to improve loading times by moving helper function to util.ts
